pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cftc-pipeline"
version = "0.1.0"
description = "End-to-end CFTC public comment ingestion and analysis pipeline"
requires-python = ">=3.11"
dependencies = [
    # DB / ORM
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    # HTTP / scraping
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "tenacity>=8.2",
    # Text extraction
    "pymupdf>=1.24",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "chardet>=5.2",
    # Deduplication
    "datasketch>=1.6",
    # LLM
    "anthropic>=0.25",
    # Embeddings / clustering
    "sentence-transformers>=3.0",
    "hdbscan>=0.8",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    # UI
    "streamlit>=1.35",
    # Report
    "jinja2>=3.1",
    "markdown>=3.6",
    # Utils
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "click>=8.1",
    "rich>=13.7",
    "tqdm>=4.66",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "responses>=0.25",
    "factory-boy>=3.3",
    "black>=24.4",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
cftc = "cftc_pipeline.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["cftc_pipeline*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
